requirements.txt
numpy == 1.18.1
PyYAML == 5.2
fire == 0.2.1
imgaug == 0.2.6
opencv_python == 4.2.0.32
pyjavaproperties == 0.7
scikit_learn == 0.20.2
scipy == 1.1.0
torch == 1.2.0
tqdm == 4.24.0
tensorboard == 1.14.0
torchvision == 0.4.0
